BUGFIX: switched to DestroyImmediate from Destroy for destructive entity operations.
Oh boy Unity at it again. So If I remove a component with Destroy, it is delayed until the end of the frame. So our filters still see that component existing there. This is not good. To fix that, we need to actually destroy the damn thing when we want, therefore I am using DestroyImmediate.